# src/docs/swagger.yaml
definitions:
...
  CarSaleAd-Web-Api_api_dto.GetOtpRequest:
    properties:
      mobileNumber:
        maxLength: 11
        minLength: 11
        type: string
    required:
    - mobileNumber
    type: object
  CarSaleAd-Web-Api_api_dto.LoginByUsernameRequest:
    properties:
      password:
        minLength: 6
        type: string
      username:
        minLength: 5
        type: string
    required:
    - password
    - username
    type: object
...
  CarSaleAd-Web-Api_api_dto.RegisterLoginByMobileRequest:
    properties:
      mobileNumber:
        maxLength: 11
        minLength: 11
        type: string
      otp:
        maxLength: 6
        minLength: 6
        type: string
    required:
    - mobileNumber
    - otp
    type: object
  CarSaleAd-Web-Api_api_dto.RegisterUserByUsernameRequest:
    properties:
      email:
        minLength: 6
        type: string
      firstName:
        minLength: 3
        type: string
      lastName:
        minLength: 6
        type: string
      password:
        minLength: 6
        type: string
      username:
        minLength: 6
        type: string
    required:
    - firstName
    - lastName
    - password
    - username
    type: object
...
  CarSaleAd-Web-Api_api_helper.BaseHttpResponse:
    properties:
      error: {}
      result: {}
      resultCode:
        $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.ResultCode'
      success:
        type: boolean
      validationErrors:
        items:
          $ref: '#/definitions/CarSaleAd-Web-Api_api_validations.ValidationError'
        type: array
    type: object
  CarSaleAd-Web-Api_api_helper.ResultCode:
    enum:
    - 0
    - 40001
    - 40101
    - 40301
    - 40401
    - 42901
    - 42902
    - 50001
    - 50002
    type: integer
    x-enum-varnames:
    - Success
    - ValidationError
    - AuthError
    - ForbiddenError
    - NotFoundError
    - LimiterError
    - OtpLimiterError
    - CustomRecovery
    - InternalError
  CarSaleAd-Web-Api_api_validations.ValidationError:
    properties:
      message:
        type: string
      property:
        type: string
      tag:
        type: string
      value:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /v1/users/login-by-mobile:
    post:
      consumes:
      - application/json
      description: RegisterLoginByMobileNumber
      parameters:
      - description: RegisterLoginByMobileRequest
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/CarSaleAd-Web-Api_api_dto.RegisterLoginByMobileRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Done
          schema:
            $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
        "409":
          description: Failed
          schema:
            $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
      summary: RegisterLoginByMobileNumber
      tags:
      - User
  /v1/users/login-by-username:
    post:
      consumes:
      - application/json
      description: LoginByUsername
      parameters:
      - description: LoginByUsernameRequest
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/CarSaleAd-Web-Api_api_dto.LoginByUsernameRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Done
          schema:
            $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
        "409":
          description: Failed
          schema:
            $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
      summary: LoginByUsername
      tags:
      - User
  /v1/users/register-by-username:
    post:
      consumes:
      - application/json
      description: RegisterByUsername
      parameters:
      - description: RegisterUserByUsernameRequest
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/CarSaleAd-Web-Api_api_dto.RegisterUserByUsernameRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Done
          schema:
            $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
        "409":
          description: Failed
          schema:
            $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
      summary: RegisterByUsername
      tags:
      - User
  /v1/users/send-otp:
    post:
      consumes:
      - application/json
      description: Test SendOtp
      parameters:
      - description: GetOtpRequest
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/CarSaleAd-Web-Api_api_dto.GetOtpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Done
          schema:
            $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
        "409":
          description: Failed
          schema:
            $ref: '#/definitions/CarSaleAd-Web-Api_api_helper.BaseHttpResponse'
      summary: SendOtp to user
      tags:
      - User
securityDefinitions:
  AuthBearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"